pyproject.toml
[project]
name = "bitthumb-cli"
version = "0.1.0"
description = "빗썸 API 이벤트 스크립트"
readme = "README.md"
authors = [{ name = "munawiki" }]
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "PyJWT>=2.8",
  "python-dotenv>=1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-mock>=3.14",
]

[project.scripts]
bitthumb-cli = "bitthumb_cli.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
